app/core/config.py
import os
from pydantic_settings import BaseSettings 
from dotenv import load_dotenv

load_dotenv()

class Settings(BaseSettings):
    """
    Application settings and environment variables.
    """
    PROJECT_NAME: str = os.getenv("PROJECT_NAME", "CreddotOne API")
    DATABASE_URL: str = os.getenv("DATABASE_URL")
    SECRET_KEY: str = os.getenv("SECRET_KEY")

    class Config:
        case_sensitive = True

settings = Settings()